requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pytesseract
requests
pillow
pydantic
ultralytics